requirements.txt
flask
segno
pillow
